Fixed Bug
Nvm Jordan, I fixed the bug.. I'm following the tutorial. It had me
assign an accel_x value in order to move the player, but for that to
work you need to set the friction. However now Ortho slides on
everything like ice.. See if you can find a way around this.