Vision Everyone Look
I've included a pdf of my vision for the first level (not finished yet),
I'll be working on transferring things as assets into the game, I was
hoping that the well would be an unlockable object after getting a key?
Is this too hard Jordan? I know Frogatto has unlockable doors and keys.